Add test file KeysConverterTests.LocalizationTests
